requirements.txt
# Core Framework
langchain
langgraph
langchain-openai
langchain-community
langchain-experimental
langchain-huggingface

# Vector Store
faiss-cpu

# Web API
fastapi
uvicorn
langserve[all]
python-multipart
httpx
requests

# Data Processing & Document Loading
pypdf
fpdf
docx2txt
openpyxl
pandas
addict
easydict
unstructured

# Search
duckduckgo-search

# Utils
python-dotenv
pydantic
sentence-transformers
transformers>=4.46.3
qwen-vl-utils
einops
timm
pillow
pdf2image
torch
torchvision
matplotlib
tiktoken

# UI
streamlit
mysql-connector-python
SQLAlchemy>=2.0.0
